Conditional agreement for more than 2 categories

Conditional agreement when there are more than two categories (averages over discordant cells to correct for random rater combinations). INCLUDE FORMULAS.

conditional.agreement(data, ...)

Arguments

data

A data matrix or table with equal number of columns and rows.

...

options for sumtable

Value

conditionaltable is a table with conditional agreement proportions. On the diagonal the specific agreement proportions for each category are displayed.

Examples

df <- data.frame(r1=factor(c(1,2,2,0,3,3,1,0,3,0,2,2,0,3,1)),
                 r2=factor(c(1,1,1,0,3,3,1,0,1,0,2,2,0,2,1)),
                 r3=factor(c(1,1,1,3,3,2,1,0,1,0,2,2,0,3,1)),
                 r4=factor(c(1,2,1,0,3,3,1,0,3,0,2,2,0,2,1)))
conditional.agreement(df)
#>   prevalence proportion     0     1     2     3
#> 0       22.5  0.2500000 0.933 0.000 0.000 0.067
#> 1       28.5  0.3166667 0.000 0.807 0.123 0.070
#> 2       21.0  0.2333333 0.000 0.167 0.667 0.167
#> 3       18.0  0.2000000 0.083 0.111 0.194 0.611