ICC consistency

The intraclass correlations (ICC) of consistency for rater reliability using the variance estimates from a linear mixed model. The function returns the ICC, standard error of measurment (sem) and confidence intervals for ICC.

Usage

icc_consistency(data, cols = colnames(data), alpha = 0.05, twoway = FALSE)

Arguments

data

data.frame with a column for each observer/rater and a row per rated subject.

cols

character vector with the column names to be used as observers. Default is `cols = colnames(data)`.

alpha

confidence interval level, default `alpha = 0.05`.

twoway

logical indicator if the variance components are estimated from the two-way model default: `twoway = FALSE`.

Value

list

Details

The ICC type consistency is the variance between the subjects divided by the sum of the subject variance and the residual variance. The subject variance and error variance are adjusted for the fixed rater effect, accordingly the rater variance is not used to calculate the ICC. The ICC for consistency generalizes only to the fixed set of raters in the data (Shrout & Fleiss, 1979). The `icc_model()` function is used to compute the variances. This is a `lmer` model with a random slope for the subjects as well as for the raters. The sem is the square root of the error variance. The confidence are computed with the exact F method. F = (k * subject variance + error variance)/ error variance, with df1 = n - 1 and df2 = (n - 1) * (k - 1) (Shrout & Fleiss, 1979).

References

Fleiss, J. L., & Shrout, P. E. Approximate interval estimation for a certain intraclass correlation coefficient. Psychometrika, 1978, 43, 259-262.

Author

Iris Eekhout