simoutput

Output results for the simulation for different icc methods under various conditions to investigate the sample size and number of raters requirements for ICC and SEM calculations.

simoutput

Format

A data frame with 2280 rows and 22 variables:

n

condition parameter: sample size

cor

condition parameter: raw correlation between raters

k

condition parameter: number of raters

variance

condition parameter: variance between patients

deviation

condition parameter: k raters with systematic deviation

method

condition parameter: icc method

set

condition parameter: simultion iteration (averaged)

icc

output parameter: average icc

lower

output parameter: average lower bound of CI for icc

upper

output parameter: average upper bound of CI for icc

sem

output parameter: average sem

varpat

output parameter: variance between patients

varobs

output parameter: variance between raters

varerr

output parameter: residual variance

Ticc

empirical population parameter: icc

Tsem

empirical population parameter: sem

bias_icc

result parameter: average bias for icc

bias_sem

result parameter: average bias for sem

mse_icc

result parameter: mean squared error for icc

mse_sem

result parameter: mean squared error for sem

cov_icc

result parameter: coverage of CI for icc

width_icc

result parameter: width of CI for icc

width_sem

result parameter: width of CI for sem

Source

data-raw/R/simoutput.R