Specific agreement-deprecated

specific agreement (averages over discordant cells to correct for random rater combinations). INCLUDE FORMULAS. When there are 2 categories, this is equal to the postive/negative agreement. When there are more than two categories, one can either look at the agreement for one category versus the others, for example very satisfied verus rest or for one category versus one specific other categorie, for example very satisfies versus not satisfied.

specific.agreement(data, cat1, cat2, ...)

Arguments

data

A data matrix or table with equal number of columns and rows. Or a data frame that contains the scores for each rater in each column.

cat1

A character indicating the category for which specific agreement should be obtained.

cat2

A character indicating the category to which the specific agreement should be compared, if left empty all other categories are used.

...

options for sumtable

Value

An S3 object containing the proportion of specific agreement.

See also