breast

For the data example we used data from a study by Dikmans et al (2017). This data is based on photographs of breasts of 50 women after breast reconstruction. The photographs are independently scored by 5 surgeons, the patient, and three mothers. They each rated the quality of the reconstruction on a 5 point ordinal scale with the verbal anchors on the left side ‘very dissatisfied’ on the left end and on the right end ‘very satisfied’ on the right end. They specifically rated the volume, shape, symmetry, scars and nipple.

breast

Format

A data frame with 50 rows and 55 variables:

Qnumber

id number

Patient_volume

Rate for volume of breast on 5pt likert by patient

Patient_shape

Rate for shape of breast on 5pt likert by patient

Patient_symmetry

Rate for symmetry of breast on 5pt likert by patient

Patient_scars

Rate for scars of breast on 5pt likert by patient

Patient_nipple

Rate for nipple of breast on 5pt likert by patient

Patient_score

Mean score breast satisfaction by patient

PCH1_volume

Rate for volume of breast on 5pt likert by surgeon1

PCH1_shape

Rate for shape of breast on 5pt likert by surgeon1

PCH1_symmetry

Rate for symmetry of breast on 5pt likert by surgeon1

PCH1_scars

Rate for scars of breast on 5pt likert by surgeon1

PCH1_nipple

Rate for nipple of breast on 5pt likert by surgeon1

PCH1_score

Mean score breast satisfaction by surgeon1

PCH2_volume

Rate for volume of breast on 5pt likert by surgeon2

PCH2_shape

Rate for shape of breast on 5pt likert by surgeon2

PCH2_symmetry

Rate for symmetry of breast on 5pt likert by surgeon2

PCH2_scars

Rate for scars of breast on 5pt likert by surgeon2

PCH2_nipple

Rate for nipple of breast on 5pt likert by surgeon2

PCH2_score

Mean score breast satisfaction by surgeon2

PCH3_volume

Rate for volume of breast on 5pt likert by surgeon3

PCH3_shape

Rate for shape of breast on 5pt likert by surgeon3

PCH3_symmetry

Rate for symmetry of breast on 5pt likert by surgeon3

PCH3_scars

Rate for scars of breast on 5pt likert by surgeon3

PCH3_nipple

Rate for nipple of breast on 5pt likert by surgeon3

PCH3_score

Mean score breast satisfaction by surgeon3

PCH4_volume

Rate for volume of breast on 5pt likert by surgeon4

PCH4_shape

Rate for shape of breast on 5pt likert by surgeon4

PCH4_symmetry

Rate for symmetry of breast on 5pt likert by surgeon4

PCH4_scars

Rate for scars of breast on 5pt likert by surgeon4

PCH4_nipple

Rate for nipple of breast on 5pt likert by surgeon4

PCH4_score

Mean score breast satisfaction by surgeon4

PCH5_volume

Rate for volume of breast on 5pt likert by surgeon5

PCH5_shape

Rate for shape of breast on 5pt likert by surgeon5

PCH5_symmetry

Rate for symmetry of breast on 5pt likert by surgeon5

PCH5_scars

Rate for scars of breast on 5pt likert by surgeon5

PCH5_nipple

Rate for nipple of breast on 5pt likert by surgeon5

PCH5_score

Mean score breast satisfaction by surgeon5

Mam1_volume

Rate for volume of breast on 5pt likert by mother1

Mam1_shape

Rate for shape of breast on 5pt likert by mother1

Mam1_symmetry

Rate for symmetry of breast on 5pt likert by mother1

Mam1_scars

Rate for scars of breast on 5pt likert by mother1

Mam1_nipple

Rate for nipple of breast on 5pt likert by mother1

Mam1_score

Mean score breast satisfaction by mother1

Mam2_volume

Rate for volume of breast on 5pt likert by mother2

Mam2_shape

Rate for shape of breast on 5pt likert by mother2

Mam2_symmetry

Rate for symmetry of breast on 5pt likert by mother2

Mam2_scars

Rate for scars of breast on 5pt likert by mother2

Mam2_nipple

Rate for nipple of breast on 5pt likert by mother2

Mam2_score

Mean score breast satisfaction by mother2

Mam3_volume

Rate for volume of breast on 5pt likert by mother3

Mam3_shape

Rate for shape of breast on 5pt likert by mother3

Mam3_symmetry

Rate for symmetry of breast on 5pt likert by mother3

Mam3_scars

Rate for scars of breast on 5pt likert by mother3

Mam3_nipple

Rate for nipple of breast on 5pt likert by mother3

Mam3_score

Mean score breast satisfaction by mother3

@source data-raw/R/breast.R